requirements.txt
numpy
pandas
Pillow
matplotlib
torch
torchvision
wandb
jupyter